👷 port per-API performance tests to benchmark harness
Move CPU and memory measurements for individual RUM APIs (addAction,
addError, addTiming, setGlobalContextProperty, startView,
start/stopSessionReplayRecording) from the legacy synthetics-driven
scripts to the Playwright benchmark harness in test/performance.

- Add apiCalls.scenario.ts driving each API via page.evaluate under the
  existing CPU/memory profilers; reuses the instrumentation-overhead app
- Allow createBenchmarkTest to opt into a subset of SDK configurations
- Resolve apiCall_* scenarios to the instrumentation-overhead app in the
  performance server
- Drop scripts/performance/{cpu,memory}Performance.ts, constants.ts, and
  the related PR comment sections — the benchmark job now reports these
  metrics directly
- Rename the test-performance GitLab job to bundle-size to reflect its
  remaining responsibility

👷 widen apiCall benchmark configurations to rum/rum_replay/rum_profiling

Allows the per-API scenarios to surface how enabling session replay or
profiling affects per-call SDK overhead, instead of only measuring the
vanilla `rum` config.

🐛 sample heap after the apiCall workload, not before

`startMemoryProfiling.stopMemoryProfiling()` discards the final
sampling profile and only takes the median of samples captured during
explicit `takeMeasurements()` calls. With the call placed before the
workload, the per-API `browser_sdk.benchmark.memory` metric was just
the post-load baseline.

Move the call to after the workload via a `runApiBenchmark` helper so
the pattern is shared across scenarios.

👷 move per-API CPU/memory tests to a scheduled api-performance-benchmark job

The new `apiCall_*` scenarios under `test/performance/` produced numbers
that aren't order-of-magnitude comparable to the old per-action metrics
(scenario-total vs per-call methodology), so switch approach: keep the
original synthetics-driven CPU test and puppeteer-driven memory test,
strip the PR-comparison parts, and run them in a new scheduled GitLab
job gated by `TARGET_TASK_NAME == "performance-benchmark-scheduled"`.

- Add `scripts/api-performance/` with trimmed CPU + memory scripts and
  their own copy of `reportToDatadog` (lint forbids cross-script
  protected-directory imports)
- Add the `api-performance-benchmark` job in `.gitlab-ci.yml`
- Revert the per-API scenarios and the `configurations` option I added
  to `createBenchmarkTest`; revert the `apiCall_*` fallback in
  `server.ts`

👷 log per-API CPU and memory tables to the CI console

After the synthetics CPU test finishes, query the just-reported metrics
from Datadog and print a small table. For memory, print the same shape
of table inline from the values the script already computed. No
comparison against a base commit — just current results.
